把所有blogId 由Long 改为BigInteger，干！这个错以后不能犯了，一定要在写数据库的时候就确立下来 字段的类型

+++ src/main/java/com/kaige/service/BlogService.java
@@ -20,7 +20,7 @@ public interface BlogService {
 
     List<Blog> getSearchBlogListByQueryAndPublished(String trim);
 
-    Boolean getCommentEnabledByBlogId(Long blogId);
+    Boolean getCommentEnabledByBlogId(BigInteger blogId);
 
-    Boolean getPublishedByBlogId(Long blogId);
+    Boolean getPublishedByBlogId(BigInteger blogId);
 }
